requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

accelerate
av
bitsandbytes
build
cmake
compressed-tensors
ninja
numpy
pillow
protobuf
pydantic
requests
streamlit
transformers
transformers[sentencepiece]
triton-windows
wheel